Refactor healpix dataset from catalog
